django project/aps_ce_backend/requirements.txt
asgiref==3.5.2
blis==0.7.8
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.5
confection==0.0.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
fonttools==4.37.1
gensim==3.6.0
greenlet==1.1.3
idna==2.10
Jinja2==2.11.3
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib==3.2.2
murmurhash==1.0.8
mysql==0.0.3
mysqlclient==2.1.1
nltk==3.7
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pandasql==0.7.3
pathy==0.6.2
Pillow==7.1.2
preshed==3.0.7
pydantic==1.9.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
regex==2022.6.2
requests==2.23.0
scikit-learn==1.0.2
scikit-surprise==1.1.1
scipy==1.7.3
six==1.15.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.41
sqlparse==0.4.3
srsly==2.4.4
thinc==8.1.3
threadpoolctl==3.1.0
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.1.1
tzdata==2022.2
urllib3==1.24.3
wasabi==0.10.1
pickle5==0.0.11
pickleshare==0.7.5
django-picklefield==3.1